+++ library/docs/changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `setOptions(options)` overload which accepts collection instead of varargs.
+
 ## [1.0.0-rc2] - 2021-12-28
 
 ### Changed